Add async LED writer to offload ALSA writes from hot paths
Introduce LedWriter with dedicated ALSA ctl handle and worker thread,
removing synchronous snd_ctl_elem_write calls from the MIDI callback and
evdev loop. LED updates now enqueue instantly and apply asynchronously,
eliminating ALSA contention as a source of input latency.

Also included:
- Signal-driven shutdown via atomic flag checked in the poll loop
- Shift keys switched from toggle to momentary behavior
- EvDevEvent type dispatch changed from string map to integer switch
- Jog touch edge detection to suppress duplicate touch events
- MIDI output queue changed to deque with 256-message depth limit
  and oldest-jog-message eviction under pressure
- Utility LED helpers take LedWriter* with static cache for VU meters
- Fix spdlog default logger segfault (was dropped before LedWriter init)

+++ includes/Knob.h
@@ -5,6 +5,7 @@
 #include <string>
 #include <map>
 #include <chrono>
+#include <cstdint>
 // --------------------------
 #include "MidiEventOut.h"
 #include "MidiHelper.h"
@@ -28,7 +29,7 @@ class Knob
     int value;
     int counter;
     int prev_control_value;
-    int updated;
+    int64_t updated;
     static map<int, Knob *> knob_mapping;
     int handle_event(RtMidiOut *midi_out_port,
                      bool shift_ch1,
