+++ includes/MidiEventOut.h
@@ -6,6 +6,7 @@
 #include <map>
 #include <unistd.h>
 #include <chrono>
+#include <cstdint>
 // --------------------------
 
 using namespace std;
@@ -34,7 +35,7 @@ public:
     unsigned char tgl_on_shf_on_status_byte;
     unsigned char tgl_on_shf_on_channel_byte;
     static map<int, MidiEventOut *> midi_mapping;
-    static int get_time();
+    static int64_t get_time();
 };
 
 #endif //TRAKTOR_KONTROL_S4_MK1_DRIVER_LINUX_RTMIDIEVENT_H
